requirements.txt
gradio>=4.0.0
huggingface_hub>=0.20.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
